(PC-19910): refactor(share): move useShareVenue in features/share
