+++ src/features/venue/components/VenueHeader/VenueHeader.native.test.tsx
@@ -4,7 +4,7 @@ import { Animated, Share, Platform } from 'react-native'
 import { mockGoBack } from 'features/navigation/__mocks__/useGoBack'
 import { VenueHeader } from 'features/venue/components/VenueHeader/VenueHeader'
 import { venueResponseSnap } from 'features/venue/fixtures/venueResponseSnap'
-import { getVenueUrl } from 'features/venue/helpers/useShareVenue'
+import { getVenueUrl } from 'features/share/helpers/useShareVenue'
 import { analytics } from 'libs/firebase/analytics'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { act, fireEvent, render } from 'tests/utils'
