+++ src/features/venue/components/VenueHeader/VenueHeader.tsx
@@ -5,8 +5,8 @@ import styled, { useTheme } from 'styled-components/native'
 
 import { getTabNavConfig } from 'features/navigation/TabBar/helpers'
 import { useGoBack } from 'features/navigation/useGoBack'
+import { useShareVenue } from 'features/share/helpers/useShareVenue'
 import { WebShareModal } from 'features/share/pages/WebShareModal'
-import { useShareVenue } from 'features/venue/helpers/useShareVenue'
 import { analytics } from 'libs/firebase/analytics'
 import { useWhiteStatusBar } from 'libs/hooks/useWhiteStatusBar'
 import { getAnimationState } from 'ui/animations/helpers/getAnimationState'
